fix: unnecessary stylesheet_link_tag for tailwind
In Rails 8, propshaft detects `:app` and loads all the css files under
app/assets/stylesheets, so we may not need this explicit tag for tailwind.

On installation, omit the tailwind stylesheet link tag if `:app` is
already being passed to stylesheet_link_tag in the layout.

On upgrade, remove the tailwind stylesheet link tag if `:app` is
already being passed to stylesheet_link_tag in the layout.